Split RL agents into separate modules
